# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.5.1",
    "title": "Supermercado SQ - Devhub API Documentation",
    "description": "Devhub's API documentation for SupermercadoSQ, developed during Atlantico Academy Bootcamp"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/restriction/": {
      "post": {
        "tags": [
          "Restrição"
        ],
        "summary": "Cria uma restrição",
        "description": "Cria uma restrição com base no nome",
        "operationId": "postRestriction",
        "parameters": [
          {
            "name": "nome_restricao",
            "in": "body",
            "description": "Nome de uma restrição",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome_restricao": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          }
        }
      },
      "get": {
        "tags": [
          "Restrição"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          }
        }
      }
    },
...
  }
}